@@ -1,327 +1,356 @@
-import React from 'react';
-import { ICPIcon, OGYIcon } from '@dapp/common-assets';
-import { AuthContext, getCanisterId } from '@dapp/features-authentication'
-import { NatPrice } from '@dapp/features-components';
-import ExpandLess from '@mui/icons-material/ExpandLess';
-import ExpandMore from '@mui/icons-material/ExpandMore';
-import FilterListIcon from '@mui/icons-material/FilterList';
-import GridViewIcon from '@mui/icons-material/GridView';
-import PaidIcon from '@mui/icons-material/Paid';
-import ViewComfyIcon from '@mui/icons-material/ViewComfy';
+import React, { useContext, useEffect, useState } from 'react';
+import { useSnackbar } from 'notistack';
+import { useDebug } from '@dapp/features-debug-provider';
+import { AuthContext, useRoute } from '@dapp/features-authentication';
+import { LoadingContainer, TokenIcon } from '@dapp/features-components';
+import { PlaceholderImage } from '@dapp/common-assets';
+import { useMarketplace } from '../../components/context';
 import {
   Card,
-  CardContent,
-  CardMedia,
-  Collapse,
-  FormControl,
+  Container,
+  Flex,
+  HR,
   Grid,
-  InputLabel,
-  List,
-  ListItem,
-  ListItemButton,
-  ListItemIcon,
-  ListItemText,
-  MenuItem,
-  Select,
-  ToggleButton,
-  ToggleButtonGroup,
-  Typography,
-} from '@mui/material';
-import Link from '@mui/material/Link';
-import TextField from '@mui/material/TextField';
-// import { useSnackbar } from 'notistack';
-import { useContext, useEffect, useState } from 'react';
+  Image,
+  SecondaryNav,
+  ShowMoreBlock,
+} from '@origyn-sa/origyn-art-ui';
+import Filter from '../../../../vault/src/pages/Vault/Filter';
+import { getNftCollectionMeta, OrigynClient } from '@origyn-sa/mintjs';
+import { Link } from 'react-router-dom';
+import { useDialog } from '@connect2ic/react';
+import styled from 'styled-components';
+import { OdcDataWithSale, parseOdcs, parseMetadata, toLargerUnit } from '@dapp/utils';
+import { Principal } from '@dfinity/principal';
 
-const SymbolWithIcon = ({ symbol }: any) =>
-  symbol === 'OGY' ? (
-    <>
-      <OGYIcon
-        style={{
-          verticalAlign: 'middle',
-          width: '20px',
-          height: '20px',
-          marginRight: '3px',
-          borderRadius: '25px',
-          backgroundColor: 'black',
-        }}
-      />{' '}
-      {symbol}
-    </>
-  ) : (
-    <>
-      <ICPIcon
-        style={{
-          verticalAlign: 'middle',
-          width: '20px',
-          height: '20px',
-          marginRight: '3px',
-          borderRadius: '25px',
-          backgroundColor: 'black',
-        }}
-      />{' '}
-      {symbol}
-    </>
-  );
+const StyledSectionTitle = styled.h2`
+  margin: 48px 24px;
+`;
 
 const Marketplace = () => {
-  const { actor } = useContext(AuthContext);
-  const [canisterId, setCanisterId] = useState("");
-  const [NFTData, setNFTData] = useState<any>();
-  const [filteredNFTs, setFilteredNFTs] = useState([]);
-  //const [isLoading, setIsLoading] = useState(true);
-  const [display, setDisplay] = useState(3);
-  const [open, setOpen] = useState(false);
-  const [openPrice, setOpenPrice] = useState(false);
-  const [onSale, setOnSale] = useState(true);
-  const [minPrice, setMinPrice] = useState<any>(0);
-  const [maxPrice, setMaxPrice] = useState<any>(0);
+  const debug = useDebug();
+  const { enqueueSnackbar } = useSnackbar() || {};
+  const { principal, actor, handleLogOut } = useContext(AuthContext);
+  const [principalId, setPrincipalId] = useState<string>();
+  const [canisterId, setCanisterId] = useState('');
+  const [isLoaded, setIsLoaded] = useState(false);
+  const [inputText, setInputText] = useState('');
+  const { open } = useDialog();
+  const { state, dispatch } = useMarketplace();
+  const { totalItems, collectionData, odcs, filter, sort, filteredOdcs } = state;
 
-  const fetchData = () => {
-    console.log(actor);
-    if (actor) {
-      //setIsLoading(true);
-      actor?.collection_nft_origyn([]).then((response) => {
-        console.log(response);
+  const logout = async () => {
+    handleLogOut();
+    fetchData();
+  };
 
-        Promise.all(
-          response?.ok?.token_ids[0]?.map((nft) => actor?.nft_origyn(nft).then((r) => r.ok)),
-        )
-          .then((data: any) => {
-            //setIsLoading(false);
-            console.log(data);
-            setNFTData(data);
-          })
-          .catch((err) => {
-            //setIsLoading(false);
-            console.log(err);
-          });
+  const fetchData = async () => {
+    if (!actor) {
+      return;
+    }
+
+    try {
+      OrigynClient.getInstance().init(true, canisterId, { actor });
+
+      // get the canister's collection metadata
+      const collMetaResp = await getNftCollectionMeta([]);
+      debug.log('return value from getNftCollectionMeta([])');
+      debug.log(JSON.stringify(collMetaResp, null, 2));
+
+      if (collMetaResp.err) {
+        // TODO: Display error
+        debug.error(collMetaResp.err);
+        return;
+      }
+
+      const collMeta = collMetaResp.ok;
+      const metadataClass = collMeta?.metadata?.[0]?.Class;
+      const collData = parseMetadata(metadataClass);
+      dispatch({ type: 'collectionData', payload: collData });
+
+      // set number of tokens
+      const tokenIds = collMeta?.token_ids?.[0] || [];
+      dispatch({ type: 'totalItems', payload: tokenIds.length });
+
+      // get a list of all digital certificates in the collection
+      const odcDataRaw = await actor?.nft_batch_origyn(tokenIds);
+      debug.log('actor?.nft_batch_origyn(ownedTokenIds)');
+      debug.log(JSON.stringify(odcDataRaw, null, 2));
+
+      if (odcDataRaw.err) {
+        // TODO: Display error
+        debug.error(odcDataRaw.err);
+        return;
+      }
+
+      // parse the digital certificate data (metadata and sale info)
+      const parsedOdcs = parseOdcs(odcDataRaw);
+      debug.log('parsed odcs');
+      debug.log(parsedOdcs);
+
+      dispatch({ type: 'odcs', payload: parsedOdcs });
+      dispatch({ type: 'filteredOdcs', payload: parsedOdcs });
+    } catch (err) {
+      debug.error(err);
+      enqueueSnackbar(err?.message || err, {
+        variant: 'error',
+        anchorOrigin: {
+          vertical: 'top',
+          horizontal: 'right',
+        },
       });
+    } finally {
+      setIsLoaded(true);
     }
   };
 
   useEffect(() => {
-    getCanisterId().then((r) => {
-      setCanisterId(r);
-    });
-  }, [])
+    document.title = 'Origyn Marketplace';
+
+    const run = async () => {
+      const route = await useRoute();
+      setCanisterId(route.canisterId);
+    };
+
+    run();
+  }, []);
 
   useEffect(() => {
-    fetchData();
+    setPrincipalId(
+      !principal || principal.toText() === Principal.anonymous().toText() ? '' : principal.toText(),
+    );
+  }, [principal]);
+
+  /* Fetch data from canister when the actor reference
+   * is ready, then every 5 seconds */
+  useEffect(() => {
+    let intervalId: any;
+    if (actor) {
+      fetchData();
+      if (!intervalId) {
+        intervalId = setInterval(() => {
+          fetchData();
+        }, 5000);
+      }
+    }
+    return () => {
+      if (intervalId) {
+        clearInterval(intervalId);
+      }
+    };
   }, [actor]);
 
+  /* Apply filter and sort to list */
   useEffect(() => {
-    const tmpFiltered = [...(NFTData || [])];
-    console.log(tmpFiltered);
-    const filtered = tmpFiltered.filter((item) => {
-      if (onSale) {
-        const currentOpenAuction = item?.current_sale?.find((sale) =>
-          sale?.sale_type?.auction?.status?.hasOwnProperty('open'),
-        );
-        console.log(currentOpenAuction);
-        return currentOpenAuction;
-      }
-      return true;
-    });
-    console.log(filtered);
-    setFilteredNFTs(filtered);
-  }, [onSale, minPrice, maxPrice, NFTData]);
+    let filtered = odcs;
+
+    switch (filter) {
+      case 'onSale':
+        filtered = filtered.filter((odc) => odc.auctionOpen);
+        break;
+      case 'notOnSale':
+        filtered = filtered.filter((odc) => !odc.auctionOpen);
+        break;
+    }
+
+    switch (sort) {
+      case 'saleASC':
+        filtered = [...filtered].sort((odc1, odc2) => {
+          return Math.max(odc2.buyNow, odc2.currentBid) - Math.max(odc1.buyNow, odc1.currentBid);
+        });
+        break;
+      case 'saleDESC':
+        filtered = [...filtered].sort((odc1, odc2) => {
+          return Math.max(odc1.buyNow, odc1.currentBid) - Math.max(odc2.buyNow, odc2.currentBid);
+        });
+        break;
+    }
+
+    if (inputText?.length) {
+      filtered = filtered.filter((odc) => odc?.displayName?.toLowerCase().includes(inputText));
+    }
+
+    dispatch({ type: 'filteredOdcs', payload: filtered });
+  }, [filter, sort, inputText, odcs]);
 
   return (
-    <div>
-      <Grid
-        container
-        sx={{
-          borderTop: '1px solid #aaaaaa',
-          minHeight: 'calc(100vh - 64px)',
-        }}
-      >
-        <Grid container item xs={12} md={3} sx={{ borderRight: '1px solid #aaaaaa' }}>
-          <List sx={{ width: '100%' }}>
-            <ListItem disablePadding>
-              <ListItemButton>
-                <ListItemIcon>
-                  <FilterListIcon />
-                </ListItemIcon>
-                <ListItemText primary="Filters" />
-              </ListItemButton>
-            </ListItem>
-            <ListItem disablePadding onClick={() => setOpenPrice(!openPrice)}>
-              <ListItemButton>
-                <ListItemIcon>
-                  <PaidIcon />
-                </ListItemIcon>
-                <ListItemText primary="Price" />
-                {openPrice ? <ExpandLess /> : <ExpandMore />}
-              </ListItemButton>
-            </ListItem>
-
-            <Collapse in={openPrice} timeout="auto" unmountOnExit>
-              <ListItem>
-                <Grid container spacing={2}>
-                  <Grid item xs={6}>
-                    <TextField
-                      id="outlined-basic"
-                      label="Min Price"
-                      variant="outlined"
-                      value={minPrice}
-                      onChange={(e) => setMinPrice(e.target.value)}
-                    />
-                  </Grid>
-                  <Grid item xs={6}>
-                    <TextField
-                      id="outlined-basic"
-                      label="Max Price"
-                      variant="outlined"
-                      value={maxPrice}
-                      onChange={(e) => setMaxPrice(e.target.value)}
-                    />
-                  </Grid>
-                </Grid>
-              </ListItem>
-            </Collapse>
-            <ListItem disablePadding onClick={() => setOpen(!open)}>
-              <ListItemButton>
-                <ListItemIcon>
-                  <FilterListIcon />
-                </ListItemIcon>
-                <ListItemText primary="Sale Token" />
-                {open ? <ExpandLess /> : <ExpandMore />}
-              </ListItemButton>
-            </ListItem>
-            <Collapse in={open} timeout="auto" unmountOnExit>
-              <List component="div" disablePadding>
-                <ListItemButton sx={{ pl: 2 }}>
-                  <ListItemIcon>
-                    <OGYIcon />
-                  </ListItemIcon>
-                  <ListItemText primary="OGY" />
-                </ListItemButton>
-                <ListItemButton sx={{ pl: 2 }}>
-                  <ListItemIcon>
-                    <ICPIcon />
-                  </ListItemIcon>
-                  <ListItemText primary="ICP" />
-                </ListItemButton>
-              </List>
-            </Collapse>
-          </List>
-        </Grid>
-        <Grid item xs={12} md={9} sx={{ padding: '10px' }}>
-          <Grid container spacing={2}>
-            <Grid item>
-              <ToggleButtonGroup size="large">
-                <ToggleButton value="left" key="left" onClick={() => setDisplay(3)}>
-                  <ViewComfyIcon />
-                </ToggleButton>
-                <ToggleButton value="left" key="left" onClick={() => setDisplay(2)}>
-                  <GridViewIcon />
-                </ToggleButton>
-              </ToggleButtonGroup>
-            </Grid>
-            <Grid item>
-              <FormControl variant="standard" sx={{ m: 1, minWidth: 120 }}>
-                <InputLabel id="demo-simple-select-label">Showing</InputLabel>
-                <Select
-                  labelId="demo-simple-select-label"
-                  id="demo-simple-select"
-                  label="Age"
-                  value={onSale ? '0' : '1'}
-                  onChange={(e) => setOnSale(e.target.value === '0')}
-                >
-                  <MenuItem value="0">On sale</MenuItem>
-                  <MenuItem value="1">All NFTs</MenuItem>
-                </Select>
-              </FormControl>
-            </Grid>
-            <Grid item>
-              <FormControl variant="standard" sx={{ m: 1, minWidth: 120 }}>
-                <InputLabel id="demo-simple-select-label">Sort By</InputLabel>
-                <Select labelId="demo-simple-select-label" id="demo-simple-select" label="Age">
-                  <MenuItem value="0">Price: Low to High</MenuItem>
-                  <MenuItem value="1">Price: High to Low</MenuItem>
-                </Select>
-              </FormControl>
-            </Grid>
-          </Grid>
-          <Grid container spacing={2}>
-            {filteredNFTs?.map((nft, index) => {
-              const nftID = nft.metadata.Class.find(({ name }) => name === 'id').value.Text;
-              return (
-                <Grid item xs={12} md={display} key={`${nft}+${index}`}>
-                  <Link href={`#/${nftID}`}>
-                    <Card variant="outlined">
-                      <CardMedia
-                        component="img"
-                        image={`https://${canisterId}.raw.ic0.app/-/${nftID}/preview`}
-                        alt={nftID}
-                      />
-                      <CardContent>
-                        <Typography variant="body2" color="text.secondary">
-                          NFT id: {nftID} <br />
-                        </Typography>
-                        <Typography>
-                          Token:{' '}
-                          <SymbolWithIcon
-                            symbol={
-                              nft?.current_sale[0]?.sale_type?.auction?.config?.auction?.token?.ic
-                                ?.symbol
-                            }
+    <Flex fullWidth padding="0" flexFlow="column">
+      <SecondaryNav
+        title="Marketplace"
+        tabs={[{ title: 'Marketplace', id: 'Marketplace' }]}
+        content={[
+          <Flex fullWidth flexFlow="column" key="marketplace-nav">
+            <StyledSectionTitle>Marketplace Dashboard</StyledSectionTitle>
+            <HR />
+            {!isLoaded ? (
+              <LoadingContainer />
+            ) : (
+              <>
+                {collectionData && (
+                  <div>
+                    <Container padding="32px">
+                      <Flex align="flex-start" gap={24}>
+                        {collectionData.hasPreviewAsset ? (
+                          <Image
+                            src={`https://prptl.io/-/${canisterId}/collection/preview`}
+                            alt="text"
+                            style={{ width: 200 }}
+                            onError={(e) => {
+                              e.currentTarget.src = PlaceholderImage;
+                            }}
                           />
-                        </Typography>
-                        <Typography>
-                          Start price:{' '}
-                          <strong>
-                            <NatPrice
-                              value={
-                                nft?.current_sale[0]?.sale_type?.auction?.config?.auction
-                                  ?.start_price
-                              }
-                            />
-                          </strong>
-                        </Typography>
-                        <Typography>
-                          Minimum step:{' '}
-                          <strong>
-                            <NatPrice
-                              value={
-                                nft?.current_sale[0]?.sale_type?.auction?.config?.auction
-                                  ?.min_increase?.amount
-                              }
-                            />
-                          </strong>
-                        </Typography>
-                        <Typography>
-                          Highest bid:{' '}
-                          <strong>
-                            <NatPrice
-                              value={nft?.current_sale[0]?.sale_type?.auction?.current_bid_amount}
-                            />
-                          </strong>
-                        </Typography>
-                        {nft?.current_sale[0]?.sale_type?.auction?.config?.auction?.buy_now
-                          ?.length > 0 && (
-                          <Typography>
-                            Buy now:{' '}
-                            <strong>
-                              <NatPrice
-                                value={
-                                  nft?.current_sale[0]?.sale_type?.auction?.config?.auction
-                                    ?.buy_now[0]
-                                }
-                              />
-                            </strong>
-                          </Typography>
+                        ) : (
+                          <Image src={PlaceholderImage} alt="text" style={{ width: 200 }} />
                         )}
-                      </CardContent>
-                    </Card>
-                  </Link>
-                </Grid>
-              );
-            })}
-          </Grid>
-        </Grid>
-      </Grid>
-    </div>
+                        <Flex flexFlow="column" justify="space-between" gap={8}>
+                          <h2>
+                            <b>{collectionData?.displayName}</b>
+                          </h2>
+                          <p>
+                            <span className="secondary_color">Created by </span>
+                            <span className="secondary_color">
+                              {collectionData.originatorPrincipalId || 'no creator name'}
+                            </span>
+                          </p>
+                          <br />
+                          <Flex>
+                            <Flex flexFlow="column">
+                              <h5>{totalItems}</h5>
+                              <p className="secondary_color">Total Items</p>
+                            </Flex>
+                          </Flex>
+                          <br />
+                          <ShowMoreBlock btnText="Read More">
+                            <p className="secondary_color">{collectionData?.description}</p>
+                          </ShowMoreBlock>
+                          <br />
+                          <br />
+                        </Flex>
+                      </Flex>
+                    </Container>
+                    <HR />
+                    <Container padding="32px">
+                      <Filter
+                        initialFilterValue={filter}
+                        initialSortValue={sort}
+                        onChangeFilter={(filterValue: string) =>
+                          dispatch({ type: 'filter', payload: filterValue })
+                        }
+                        onChangeSort={(sortValue: string) =>
+                          dispatch({ type: 'sort', payload: sortValue })
+                        }
+                        onInput={setInputText}
+                      />
+                      <br />
+                      <br />
+                      {filteredOdcs?.length > 0 ? (
+                        <>
+                          <Grid
+                            smColumns={1}
+                            mdColumns={2}
+                            lgColumns={3}
+                            xlColumns={4}
+                            columns={6}
+                            gap={20}
+                          >
+                            {filteredOdcs.map((odc: OdcDataWithSale) => {
+                              return (
+                                <Link to={`/${odc?.id}`} key={odc?.id}>
+                                  <Card
+                                    flexFlow="column"
+                                    style={{ overflow: 'hidden', height: '100%' }}
+                                  >
+                                    {odc.hasPreviewAsset ? (
+                                      <img
+                                        style={{ width: '100%' }}
+                                        src={`https://${canisterId}.raw.ic0.app/-/${odc?.id}/preview`}
+                                        alt=""
+                                        onError={(e) => {
+                                          e.currentTarget.src = PlaceholderImage;
+                                        }}
+                                      />
+                                    ) : (
+                                      <Image
+                                        src={PlaceholderImage}
+                                        alt="text"
+                                        style={{ width: 200 }}
+                                      />
+                                    )}
+                                    <Container
+                                      style={{ height: '100%' }}
+                                      size="full"
+                                      padding="16px"
+                                    >
+                                      <Flex
+                                        style={{ height: '100%' }}
+                                        justify="space-between"
+                                        flexFlow="column"
+                                        gap={32}
+                                      >
+                                        <div>
+                                          <p style={{ fontSize: '12px', color: '#9A9A9A' }}>
+                                            {collectionData?.displayName}
+                                          </p>
+                                          <p>
+                                            <b>{odc?.displayName || odc?.id}</b>
+                                          </p>
+                                        </div>
+                                        <div>
+                                          <p style={{ fontSize: '12px', color: '#9A9A9A' }}>
+                                            Status
+                                          </p>
+                                          <p>
+                                            {odc.auctionOpen ? (
+                                              odc.currentBid === 0 ? (
+                                                <>
+                                                  {toLargerUnit(
+                                                    odc.buyNow,
+                                                    Number(odc.token.decimals),
+                                                  )}{' '}
+                                                  <TokenIcon symbol={odc.tokenSymbol} />
+                                                </>
+                                              ) : (
+                                                <>
+                                                  {toLargerUnit(
+                                                    odc.currentBid,
+                                                    Number(odc.token.decimals),
+                                                  )}{' '}
+                                                  <TokenIcon symbol={odc.tokenSymbol} />
+                                                </>
+                                              )
+                                            ) : (
+                                              'No auction started'
+                                            )}
+                                          </p>
+                                        </div>
+                                      </Flex>
+                                    </Container>
+                                  </Card>
+                                </Link>
+                              );
+                            })}
+                          </Grid>
+                          <br />
+                        </>
+                      ) : (
+                        <h5>
+                          {odcs?.length === 0
+                            ? 'There are no digital certificates in this collection'
+                            : 'Your filter returned 0 digital certificates'}
+                        </h5>
+                      )}
+                    </Container>
+                  </div>
+                )}
+              </>
+            )}
+          </Flex>,
+        ]}
+        onLogOut={logout}
+        onConnect={open}
+        principal={principalId}
+      />
+    </Flex>
   );
 };
 
