@@ -1,10 +1,11 @@
-const webpack = require('../../../webpack.config')
+const webpack = require('../../../webpack.config');
 
 const appConfig = {
   name: 'marketplace',
-  openPage: '-/baycdev/-/bayc-1/-/marketplace',
-}
+  openPage: '-/brain-matters-dev/collection/-/marketplace',
+};
 module.exports = (env, argv) => {
-  const res = webpack(env, argv, appConfig)
-  return res
-}
+  const res = webpack(env, argv, appConfig);
+  res.output.publicPath = '/';
+  return res;
+};
