@@ -1,80 +1,96 @@
-import * as React from 'react';
-import Button from '@mui/material/Button';
-import Dialog from '@mui/material/Dialog';
-import DialogActions from '@mui/material/DialogActions';
-import DialogContent from '@mui/material/DialogContent';
+import React from 'react';
 import { Principal } from '@dfinity/principal';
-import DialogTitle from '@mui/material/DialogTitle';
 import { AuthContext } from '@dapp/features-authentication';
-import {
-  Backdrop,
-  CircularProgress,
-  FormControl,
-  InputLabel,
-  MenuItem,
-  Select,
-  TextField,
-  Typography,
-  Grid,
-} from '@mui/material';
 import { LocalizationProvider } from '@mui/x-date-pickers/LocalizationProvider';
 import { useSnackbar } from 'notistack';
-import { useForm } from 'react-hook-form';
-import { yupResolver } from '@hookform/resolvers/yup';
 import * as Yup from 'yup';
-import { DateTimePicker } from '@mui/x-date-pickers';
 import { AdapterMoment } from '@mui/x-date-pickers/AdapterMoment';
-import { TokenIcon } from '@dapp/features-components';
 import { useTokensContext } from '@dapp/features-tokens-provider';
+import {
+  Container,
+  Flex,
+  HR,
+  Modal,
+  TextInput,
+  DatePicker,
+  Select,
+  Button,
+} from '@origyn-sa/origyn-art-ui';
+import { LinearProgress } from '@mui/material';
 
-export function StartAuctionModal({ currentToken, open, handleClose }: any) {
+const dateNow = new Date();
+const dateTomorrow = new Date(new Date().valueOf() + 1000 * 3600 * 23);
+
+const validationSchema = Yup.object({
+  token: Yup.string().default('OGY'),
+  startPrice: Yup.number()
+    .typeError('This must be a number')
+    .nullable()
+    .typeError('This cannot be a nullable number')
+    .required('Start Price is required')
+    .default(0),
+  minIncrease: Yup.number()
+    .typeError('This must be a number')
+    .nullable()
+    .required('Minimum increase step is required')
+    .default(0),
+  reservePrice: Yup.number().typeError('This must be a number').nullable().default(0),
+  buyNowPrice: Yup.number()
+    .typeError('This must be a number')
+    .nullable()
+    .moreThan(Yup.ref('startPrice'), 'Instant buy price must be greater than the start price')
+    .default(0),
+  endDate: Yup.date()
+    .min(dateTomorrow, 'The end date needs to be at least one day after start date')
+    .default(dateNow),
+});
+
+export function StartAuctionModal({ currentToken, open, handleClose, onSuccess }: any) {
   const { actor } = React.useContext(AuthContext);
   const { enqueueSnackbar } = useSnackbar();
-  const [token, setToken] = React.useState('OGY');
+  const [errors, setErrors] = React.useState<any>({});
+  // const [tokenID, setTokenID] = useState<any>();
+  // @ts-ignore
+  const [values, setValues] = React.useState<any>(validationSchema.default());
   const [inProgress, setInProgress] = React.useState(false);
-  const [buyNowPrice, setBuyNowPrice] = React.useState<any>();
-  const [startPrice, setStartPrice] = React.useState<any>();
-  const [priceStep, setPriceStep] = React.useState<any>();
-  const [endDate, setEndDate] = React.useState<any>(new Date().toJSON().slice(0, 19));
-  const { tokens } = useTokensContext();
-
-  const handleChange = (event) => {
-    setToken(event.target.value as string);
-  };
+  const [success, setSuccess] = React.useState(false);
+  const { tokens, activeTokens } = useTokensContext();
 
   const handleStartAuction = async ({
     startPrice,
     buyNowPrice,
+    reservePrice,
     minIncrease: priceStep,
     endDate,
+    token: saleToken,
   }) => {
     setInProgress(true);
     try {
       const resp = await actor.market_transfer_nft_origyn({
-        token_id: currentToken?.Class?.find(({ name }) => name === 'id').value.Text,
+        token_id: currentToken,
         sales_config: {
           pricing: {
             auction: {
               start_price: BigInt(startPrice * 1e8),
               token: {
                 ic: {
-                  fee: BigInt(tokens[token]?.fee ?? 200000),
-                  decimals: BigInt(tokens[token]?.decimals ?? 8),
-                  canister: Principal.fromText(tokens[token]?.canisterId),
+                  fee: BigInt(tokens[saleToken]?.fee ?? 200000),
+                  decimals: BigInt(tokens[saleToken]?.decimals ?? 8),
+                  canister: Principal.fromText(tokens[saleToken]?.canisterId),
                   standard: { Ledger: null },
-                  symbol: tokens[token]?.symbol,
+                  symbol: tokens[saleToken]?.symbol,
                 },
               },
-              reserve: [],
+              reserve: [BigInt(reservePrice * 1e8)],
               start_date: BigInt(Math.floor(new Date().getTime() * 1e6)),
               min_increase: {
                 amount: BigInt(priceStep * 1e8),
               },
               allow_list: [],
               buy_now: [BigInt(buyNowPrice * 1e8)],
-              end_date: BigInt(new Date(endDate).getTime() * 1e6),
+              // end_date: BigInt(new Date(endDate).getTime() * 1e6), TODO: figure this out
               ending: {
-                date: BigInt(new Date(endDate).getTime() * 1e6),
+                date: BigInt(endDate.getTime() * 1e6),
               },
             },
           },
@@ -82,16 +98,8 @@ export function StartAuctionModal({ currentToken, open, handleClose }: any) {
           escrow_receipt: [],
         },
       });
-      if (resp.ok) {
-        enqueueSnackbar('Your auction has been started.', {
-          variant: 'success',
-          anchorOrigin: {
-            vertical: 'top',
-            horizontal: 'right',
-          },
-        });
-        handleClose(true);
-      } else {
+      console.log(resp);
+      if ('err' in resp) {
         enqueueSnackbar('There was an error when starting your auction.', {
           variant: 'error',
           anchorOrigin: {
@@ -99,9 +107,19 @@ export function StartAuctionModal({ currentToken, open, handleClose }: any) {
             horizontal: 'right',
           },
         });
+      } else {
+        enqueueSnackbar('Your auction has been started.', {
+          variant: 'success',
+          anchorOrigin: {
+            vertical: 'top',
+            horizontal: 'right',
+          },
+        });
+        onSuccess(resp.ok);
+        setSuccess(true);
       }
     } catch (e) {
-      console.log(e);
+      console.log('this ie error', e.message);
       enqueueSnackbar('There was an error when starting your auction.', {
         variant: 'error',
         anchorOrigin: {
@@ -112,145 +130,127 @@ export function StartAuctionModal({ currentToken, open, handleClose }: any) {
     }
     setInProgress(false);
   };
-  const validationSchema = Yup.object().shape({
-    startPrice: Yup.number()
-      .typeError('This must be a number')
-      .nullable()
-      .typeError('This cannot be a nullable number'),
-    minIncrease: Yup.string().required('Minimum increase step is required'),
-    buyNowPrice: Yup.number()
-      .typeError('This must be a number')
-      .nullable()
-      .typeError('This cannot be a nullable number')
-      .moreThan(Yup.ref('startPrice'), 'Instant buy price must be greater than the start price'),
-    endDate: Yup.date().min(new Date(), 'The end date needs to be in the future!'),
-  });
 
-  const {
-    register,
-    handleSubmit,
-    formState: { errors },
-  } = useForm({
-    resolver: yupResolver(validationSchema),
-  });
-  const customSubmit = (data) => {
-    handleStartAuction(data);
+  const getValidationErrors = (err) => {
+    const validationErrors = {};
+
+    err.inner.forEach((error) => {
+      if (error.path) {
+        validationErrors[error.path] = error.message;
+      }
+    });
+
+    return validationErrors;
   };
+  const handleSubmit = (e: any) => {
+    e.preventDefault();
+    validationSchema
+      .validate(values, { abortEarly: false })
+      .then(() => {
+        handleStartAuction(values);
+      })
+      .catch(function (e) {
+        const errs = getValidationErrors(e);
+        setErrors(errs);
+      });
+  };
+  const onChange = (e?: any, name?: string, value?: any) => {
+    setErrors({ ...errors, [name || e.target.name]: undefined });
+    setValues({ ...values, [name || e.target.name]: value || e.target.value });
+  };
+
   return (
     <div>
-      <Dialog
-        open={open}
-        onClose={() => handleClose(false)}
-        aria-labelledby="alert-dialog-title"
-        aria-describedby="alert-dialog-description"
-      >
-        <DialogTitle id="alert-dialog-title">
-          Start auction for {currentToken?.Class?.find(({ name }) => name === 'id').value.Text}?
-        </DialogTitle>
-        <DialogContent>
-          <Grid container spacing={3} mt={2}>
-            <Grid item xs={12} sm={12}>
-              <TextField
-                required
-                label="Start Price"
-                fullWidth
-                id="startPrice"
-                variant="outlined"
-                value={startPrice}
-                onChange={(e) => setStartPrice(e.target.value)}
-                {...register('startPrice')}
-                error={!!errors.startPrice}
-              />
-              <Typography variant="inherit" color="textSecondary">
-                {errors.startPrice?.message}
-              </Typography>
-            </Grid>
-            <Grid item xs={12} sm={12}>
-              <TextField
-                required
-                label="Min Increase"
-                fullWidth
-                id="minIncrease"
-                variant="outlined"
-                value={priceStep}
-                onChange={(e) => setPriceStep(e.target.value)}
-                {...register('minIncrease')}
-                error={!!errors.minIncrease}
-              />
-              <Typography variant="inherit" color="textSecondary">
-                {errors.minIncrease?.message}
-              </Typography>
-            </Grid>
-
-            <Grid item xs={12} sm={12}>
-              <TextField
-                required
-                label="Buy Now Price"
-                fullWidth
-                id="outlined-basic"
-                variant="outlined"
-                value={buyNowPrice}
-                onChange={(e) => setBuyNowPrice(e.target.value)}
-                {...register('buyNowPrice')}
-                error={!!errors.buyNowPrice}
-              />
-              <Typography variant="inherit" color="textSecondary">
-                {errors.buyNowPrice?.message}
-              </Typography>
-            </Grid>
-
-            <Grid item xs={12} sm={12}>
-              <LocalizationProvider fullWidth dateAdapter={AdapterMoment}>
-                <DateTimePicker
-                  label="End Date"
-                  value={endDate}
-                  onChange={(newValue) => {
-                    setEndDate(newValue);
-                  }}
-                  renderInput={(params) => (
-                    <TextField {...params} {...register('endDate')} error={!!errors.endDate} />
-                  )}
-                />
-              </LocalizationProvider>
-              <Typography variant="inherit" color="textSecondary">
-                {errors.endDate?.message}
-              </Typography>
-            </Grid>
-
-            <Grid item xs={12} sm={12}>
-              <FormControl fullWidth>
-                <InputLabel id="demo-simple-select-label">Token</InputLabel>
-                <Select
-                  labelId="demo-simple-select-label"
-                  id="demo-simple-select"
-                  value={token}
-                  label="Token"
-                  onChange={handleChange}
-                >
-                  {Object.keys(tokens).map((t, index) => (
-                    <MenuItem key={`${token}+${index}`} value={t}>
-                      <TokenIcon symbol={tokens[t].icon} />
-                      {tokens[t].symbol}
-                    </MenuItem>
-                  ))}
-                </Select>
-              </FormControl>
-            </Grid>
-          </Grid>
-        </DialogContent>
-        <DialogActions>
-          <Button onClick={() => handleClose(false)}>Cancel</Button>
-          <Button onClick={handleSubmit(customSubmit)} autoFocus>
-            Start
-          </Button>
-        </DialogActions>
-        <Backdrop
-          sx={{ color: '#fff', zIndex: (theme) => theme.zIndex.drawer + 1 }}
-          open={inProgress}
-        >
-          <CircularProgress color="inherit" />
-        </Backdrop>
-      </Dialog>
+      <Modal isOpened={open} closeModal={() => handleClose(false)} size="md">
+        <Container size="full" padding="48px">
+          {success ? (
+            <>
+              <h2>Success!</h2>
+              <br />
+              <p className="secondary_color">
+                Your auction has started. Click done to view the status.
+              </p>
+              <br />
+              <Flex justify="flex-end">
+                <Button onClick={handleClose}>Done</Button>
+              </Flex>
+            </>
+          ) : (
+            <>
+              {inProgress ? (
+                <>
+                  <h2>Start an Auction in Progress</h2>
+                  <br />
+                  <LinearProgress color="secondary" />
+                </>
+              ) : (
+                <>
+                  <h2>Start an Auction</h2>
+                  <br />
+                  <Flex as="form" onSubmit={handleSubmit} action="" flexFlow="column" gap={8}>
+                    <Select
+                      label="Token"
+                      name="token"
+                      /*@ts-ignore*/
+                      selectedOption={{ label: values.token, value: values.token }}
+                      handleChange={(v) => onChange(null, 'token', v.value)}
+                      options={Object.keys(activeTokens).map((t) => ({
+                        label: activeTokens[t].symbol,
+                        value: t,
+                      }))}
+                    />
+                    <TextInput
+                      required
+                      label="Starting Price"
+                      name="startPrice"
+                      value={values.startPrice}
+                      onChange={onChange}
+                      error={errors?.startPrice}
+                    />
+                    <TextInput
+                      label="Buy Now Price"
+                      name="buyNowPrice"
+                      value={values.buyNowPrice}
+                      onChange={onChange}
+                      error={errors?.buyNowPrice}
+                    />
+                    <TextInput
+                      required
+                      label="Minimum Increase"
+                      name="minIncrease"
+                      value={values.minIncrease}
+                      onChange={onChange}
+                      error={errors?.minIncrease}
+                    />
+                    <TextInput
+                      label="Reserve Price"
+                      name="reservePrice"
+                      value={values.reservePrice}
+                      onChange={onChange}
+                      error={errors?.reservePrice}
+                    />
+                    <LocalizationProvider fullWidth dateAdapter={AdapterMoment}>
+                      <DatePicker
+                        label="Set Auction Length"
+                        name="endDate"
+                        selected={values.endDate}
+                        onChange={(d) => onChange(null, 'endDate', d)}
+                        error={errors?.endDate}
+                      />
+                    </LocalizationProvider>
+                    <br />
+                    <HR />
+                    <Flex align="center" justify="flex-end" gap={16}>
+                      <Button onClick={() => handleClose(false)}>Cancel</Button>
+                      <Button type="submit">Start</Button>
+                    </Flex>
+                  </Flex>
+                </>
+              )}
+            </>
+          )}
+        </Container>
+      </Modal>
     </div>
   );
 }
