@@ -1,75 +1,59 @@
-import { ICPIcon, OGYIcon } from '@dapp/common-assets';
-import { AuthContext, getCanisterId } from '@dapp/features-authentication'
-import { NatPrice } from '@dapp/features-components';
+/* eslint-disable react/jsx-key */
+/* eslint-disable @typescript-eslint/dot-notation */
+import { useDebug } from '@dapp/features-debug-provider';
+import { AuthContext, useRoute } from '@dapp/features-authentication';
+import { LoadingContainer, TokenIcon } from '@dapp/features-components';
+import { ConfirmSalesActionModal } from '../../modals/ConfirmSalesActionModal';
+import { StartAuctionModal } from '../../modals/StartAuctionModal';
+import { StartEscrowModal } from '../../modals/StartEscrowModal';
 import {
-  ConfirmSalesActionModal,
-  StartAuctionModal,
-  StartEscrowModal,
-} from '@dapp/features-sales-escrows';
-import { eToNumber, timeConverter } from '@dapp/utils';
-import AccessTimeIcon from '@mui/icons-material/AccessTime';
-import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
-import HistoryOutlinedIcon from '@mui/icons-material/HistoryOutlined';
-import LocalOfferOutlinedIcon from '@mui/icons-material/LocalOfferOutlined';
-import PendingIcon from '@mui/icons-material/Pending';
-import ShoppingCartOutlinedIcon from '@mui/icons-material/ShoppingCartOutlined';
+  getDiffInDays,
+  OdcData,
+  OdcDataWithSale,
+  parseMetadata,
+  parseOdc,
+  toLargerUnit,
+} from '@dapp/utils';
+import { Property } from '@dapp/common-types';
+import React, { useContext, useEffect, useState } from 'react';
+import { useParams } from 'react-router-dom';
 import {
-  Accordion,
-  AccordionDetails,
-  AccordionSummary,
+  Flex,
+  HR,
+  SecondaryNav,
   Button,
   Container,
   Grid,
-  Link,
-  Typography,
-} from '@mui/material';
-import Box from '@mui/material/Box';
-import CircularProgress from '@mui/material/CircularProgress';
-import React, { useContext, useEffect, useState } from 'react';
-import { useParams, useSearchParams } from 'react-router-dom';
+  Banner,
+  TabContent,
+  ShowMoreBlock,
+} from '@origyn-sa/origyn-art-ui';
+import { useDialog } from '@connect2ic/react';
+import { getNftCollectionMeta, OrigynClient } from '@origyn-sa/mintjs';
+import { EscrowType } from '../../modals/StartEscrowModal';
+import { Principal } from '@dfinity/principal';
+import { PlaceholderImage } from '@dapp/common-assets';
+import { OffersPanel } from './components/OffersPanel';
 
-const SymbolWithIcon = ({ symbol }: any) =>
-  symbol === 'OGY' ? (
-    <>
-      <OGYIcon
-        style={{
-          verticalAlign: 'middle',
-          width: '20px',
-          height: '20px',
-          marginRight: '3px',
-          borderRadius: '25px',
-          backgroundColor: 'black',
-        }}
-      />{' '}
-      {symbol}
-    </>
-  ) : (
-    <>
-      <ICPIcon
-        style={{
-          verticalAlign: 'middle',
-          width: '20px',
-          height: '20px',
-          marginRight: '3px',
-          borderRadius: '25px',
-          backgroundColor: 'black',
-        }}
-      />{' '}
-      {symbol}
-    </>
-  );
 export const NFTPage = () => {
-  const { principal, actor } = useContext(AuthContext);
-  const [currentNFT, setCurrentNFT] = useState<any>({});
-  const [canisterId, setCanisterId] = useState("");
+  const debug = useDebug();
+  const { principal, actor, handleLogOut } = useContext(AuthContext);
+  const [principalId, setPrincipalId] = useState<string>();
+  const [odc, setOdc] = useState<OdcDataWithSale>();
+  const [collectionData, setCollectionData] = useState<OdcData>();
   const [openAuction, setOpenAuction] = React.useState(false);
+  const [canisterId, setCanisterId] = React.useState('');
   const [dialogAction, setDialogAction] = useState<any>();
   const [openConfirmation, setOpenConfirmation] = React.useState(false);
   const [isLoading, setIsLoading] = useState(true);
-  const [searchParams] = useSearchParams();
   const [openEscrowModal, setOpenEscrowModal] = React.useState(false);
-  const [modalInitialValues, setModalInitialValues] = React.useState({});
-  const [expanded, setExpanded] = React.useState<string | false>('panel1');
+  const [escrowType, setEscrowType] = React.useState<EscrowType>();
+  const { open } = useDialog();
+
+  const logout = () => {
+    handleLogOut();
+    fetchData();
+  };
 
   const handleClickOpen = (item, modal = 'auction') => {
     if (modal === 'auction') setOpenAuction(true);
@@ -79,418 +63,372 @@ export const NFTPage = () => {
     }
   };
 
+  const handleClickOpenEsc = () => {
+    setOpenConfirmation(true);
+    setDialogAction('endSale');
+  };
+
   const handleClose = async () => {
     setOpenAuction(false);
     setOpenConfirmation(false);
   };
 
-  const handleChange = (panel: string) => (event: React.SyntheticEvent, newExpanded: boolean) => {
-    setExpanded(newExpanded ? panel : false);
-  };
   const params = useParams();
-  const currentOpenAuction = currentNFT?.current_sale?.find((sale) =>
-    sale?.sale_type?.auction?.status?.hasOwnProperty('open'),
-  );
 
-  const handleOpen = (type) => {
-    const modalInitial = {
-      nftId: params.nft_id,
-      sellerId: currentNFT?.metadata?.Class?.find(
-        ({ name }) => name === 'owner',
-      ).value.Principal.toText(),
-      priceOffer: '0',
-    };
-    if (type === 'buyNow') {
-      modalInitial.priceOffer = eToNumber(
-        parseInt(currentOpenAuction?.sale_type?.auction?.config?.auction?.buy_now[0].toString()) /
-          100_000_000,
-      );
-    } else if (type === 'bid') {
-      const startPrice = parseInt(
-        currentOpenAuction?.sale_type?.auction?.config?.auction?.start_price,
-      );
-      const min_increase = parseInt(
-        currentOpenAuction?.sale_type?.auction?.config?.auction?.min_increase?.amount,
-      );
-      const highest_bid = parseInt(currentOpenAuction?.sale_type?.auction?.current_bid_amount);
-      modalInitial.priceOffer = eToNumber(
-        highest_bid > 0
-          ? (parseInt(highest_bid.toString()) + parseInt(min_increase.toString())) / 100_000_000
-          : parseInt(startPrice.toString()) / 100_000_000,
-      );
-    }
-    setModalInitialValues(modalInitial);
+  const onOpenEscrowModal = (escrowType: EscrowType) => {
+    setEscrowType(escrowType);
     setOpenEscrowModal(true);
   };
 
   const handleCloseEscrow = async (dataChanged = false) => {
     setOpenEscrowModal(false);
     if (dataChanged) {
-      // fetchData();
+      fetchOdc();
     }
   };
 
-  useEffect(() => {
-    if (searchParams.get('nftId')) {
-      const initialParameters = searchParams.entries();
-      const params = {};
-      for (const [key, value] of initialParameters) {
-        params[key] = value;
-      }
-      setModalInitialValues(params);
-      setOpenEscrowModal(true);
-    }
+  const currentTimeInNanos = BigInt(new Date().getTime() * 1e6);
 
-    if (actor) {
-      actor
-        .nft_origyn(params.nft_id)
-        .then((r) => {
-          console.log(r);
-          setIsLoading(false);
-          setCurrentNFT(r.ok);
-        })
-        .catch(console.log);
+  const nftEndSale = odc?.auction?.end_date;
+
+  const verifyOwner = odc?.ownerPrincipalId || '';
+
+  const fetchCollection = async () => {
+    const route = await useRoute();
+    setCanisterId(route.canisterId);
+
+    OrigynClient.getInstance().init(true, canisterId, { actor });
+
+    const collMetaResp = await getNftCollectionMeta([]);
+    debug.log('return value from getNftCollectionMeta([])');
+    debug.log(JSON.stringify(collMetaResp, null, 2));
+
+    if (collMetaResp.err) {
+      setCollectionData(undefined);
+      debug.error(collMetaResp.err);
+    } else {
+      const collMeta = collMetaResp.ok;
+      const metadataClass = collMeta?.metadata?.[0]?.Class as Property[];
+      const parsedCollData = parseMetadata(metadataClass);
+      setCollectionData(parsedCollData);
     }
+  };
+
+  const fetchOdc = async () => {
+    const r: any = await actor.nft_origyn(params.nft_id);
+    debug.log('return value from actor.nft_origyn(params.nft_id)');
+    debug.log(JSON.stringify(r, null, 2));
+
+    if ('err' in r) {
+      throw new Error(Object.keys(r.err)[0]);
+    }
+    setIsLoading(false);
+
+    let parsedOdc: OdcDataWithSale = parseOdc(r['ok']);
+    debug.log('parsedOdc');
+    debug.log(parsedOdc);
+
+    setOdc(parsedOdc);
+  };
+
+  const fetchData = async () => {
+    Promise.all([fetchCollection(), fetchOdc()]);
+  };
+
+  useEffect(() => {
+    setPrincipalId(
+      !principal || principal.toText() === Principal.anonymous().toText() ? '' : principal.toText(),
+    );
+  }, [principal]);
+
+  useEffect(() => {
+    let intervalId: any;
+    if (actor) {
+      fetchData();
+      if (!intervalId) {
+        intervalId = setInterval(() => {
+          fetchData();
+        }, 5000);
+      }
+    }
+    return () => {
+      if (intervalId) {
+        clearInterval(intervalId);
+      }
+    };
   }, [actor]);
 
-  useEffect(() => {
-    getCanisterId().then((r) => {
-      setCanisterId(r);
-    });
-  }, []);
-
-  if (isLoading || !canisterId) {
-    return (
-      <Box sx={{ display: 'flex', justifyContent: 'center' }}>
-        <CircularProgress />
-      </Box>
-    );
-  }
   return (
-    <div>
-      <StartEscrowModal
-        open={openEscrowModal}
-        handleClose={handleCloseEscrow}
-        nft={currentNFT}
-        initialValues={modalInitialValues}
-      />
-      <Container sx={{ mt: 4, mb: 4 }}>
-        <Grid container spacing={4}>
-          <Grid item xs={6} md={5}>
-            <img
-              width="100%"
-              style={{ margin: '0 10px 10px 10px' }}
-              src={`https://${canisterId}.raw.ic0.app/-/${params.nft_id}/preview`}
-            />
-            <div>
-              <Accordion expanded={expanded === 'panel1'} onChange={handleChange('panel1')}>
-                <AccordionSummary aria-controls="panel1d-content" id="panel1d-header">
-                  <Typography>Description</Typography>
-                </AccordionSummary>
-                <AccordionDetails>
-                  <Typography>
-                    {currentNFT?.metadata?.Class?.find(({ name }) => name === 'description')?.value
-                      ?.Text || <b>No description provided</b>}
-                  </Typography>
-                </AccordionDetails>
-              </Accordion>
-              <Accordion expanded={expanded === 'panel2'} onChange={handleChange('panel2')}>
-                <AccordionSummary aria-controls="panel2d-content" id="panel2d-header">
-                  <Typography>Properties</Typography>
-                </AccordionSummary>
-                <AccordionDetails>
-                  <Typography>
-                    {currentNFT?.metadata?.Class?.find(({ name }) => name === '__apps')
-                      ?.value?.Array?.thawed[0].Class.find(({ name }) => name === 'data')
-                      .value.Class.map(({ name, value, index }) => (
-                        <div key={`${name}+${index}`}>
-                          {name}: <b>{Object.values(value)[0].toString()}</b>
-                        </div>
-                      )) || <b>No properties available</b>}
-                  </Typography>
-                </AccordionDetails>
-              </Accordion>
-            </div>
-          </Grid>
-          <Grid
-            item
-            xs={6}
-            md={7}
-            style={{
-              display: 'flex',
-              gap: '5px',
-              flexWrap: 'wrap',
-              flexDirection: 'column',
-            }}
-          >
-            <Typography variant="h5">{params.nft_id}</Typography>
-            <Typography>
-              Owned by{' '}
-              <Link href="#">
-                {currentNFT?.metadata?.Class?.find(
-                  ({ name }) => name === 'owner',
-                ).value.Principal.toText()}
-              </Link>
-            </Typography>
-            {currentNFT?.metadata?.Class?.find(
-              ({ name }) => name === 'owner',
-            ).value.Principal.toText() === principal?.toText() ? (
-              <div>
-                {currentOpenAuction ? (
-                  <Accordion expanded>
-                    <AccordionSummary
-                      expandIcon={<ExpandMoreIcon />}
-                      aria-controls="panel1a-content"
-                      id="panel1a-header"
-                    >
-                      <Grid container direction="row" alignItems="center">
-                        <AccessTimeIcon style={{ marginRight: 2 }} /> There is an active auction for
-                        this NFT
-                      </Grid>
-                    </AccordionSummary>
-                    <AccordionDetails>
-                      <Typography>
-                        Sale ends on{' '}
-                        <strong>
-                          {' '}
-                          {timeConverter(
-                            BigInt(parseInt(currentOpenAuction?.sale_type?.auction?.end_date)),
-                          )}
-                        </strong>
-                      </Typography>
-                      <Typography>
-                        Token:{' '}
-                        <SymbolWithIcon
-                          symbol={
-                            currentOpenAuction?.sale_type?.auction?.config?.auction?.token?.ic
-                              ?.symbol
-                          }
-                        />
-                      </Typography>
-                      <Typography>
-                        Start price:{' '}
-                        <strong>
-                          <NatPrice
-                            value={
-                              currentOpenAuction?.sale_type?.auction?.config?.auction?.start_price
-                            }
-                          />
-                        </strong>
-                      </Typography>
-                      <Typography>
-                        Minimum step:{' '}
-                        <strong>
-                          <NatPrice
-                            value={
-                              currentOpenAuction?.sale_type?.auction?.config?.auction?.min_increase
-                                ?.amount
-                            }
-                          />
-                        </strong>
-                      </Typography>
-                      <Typography>
-                        Highest bid:{' '}
-                        <strong>
-                          <NatPrice
-                            value={currentOpenAuction?.sale_type?.auction?.current_bid_amount}
-                          />
-                        </strong>
-                      </Typography>
-                      {currentOpenAuction?.sale_type?.auction?.config?.auction?.buy_now?.length >
-                        0 && (
-                        <Typography>
-                          Buy now:{' '}
-                          <strong>
-                            <NatPrice
-                              value={
-                                currentOpenAuction?.sale_type?.auction?.config?.auction?.buy_now[0]
-                              }
-                            />
-                          </strong>
-                        </Typography>
-                      )}
-                      <div style={{ display: 'flex', gap: 5, marginTop: 5 }} />
-                    </AccordionDetails>
-                  </Accordion>
+    <>
+      {odc && (
+        <Flex fullWidth padding="0" flexFlow="column">
+          <SecondaryNav
+            title="Vault"
+            tabs={[{ title: 'NFT Details', id: 'nft' }]}
+            content={[
+              <Flex fullWidth flexFlow="column">
+                {isLoading ? (
+                  <LoadingContainer />
                 ) : (
-                  <Accordion expanded>
-                    <AccordionSummary
-                      expandIcon={<ExpandMoreIcon />}
-                      aria-controls="panel1a-content"
-                      id="panel1a-header"
-                    >
-                      <Grid container direction="row" alignItems="center">
-                        <PendingIcon style={{ marginRight: 2 }} /> This NFT is not listed for a
-                        public sale.
-                      </Grid>
-                    </AccordionSummary>
-                    <AccordionDetails>
-                      You can start an Auction for this NFT.
-                      <div style={{ display: 'flex', gap: 5, marginTop: 5 }}>
-                        <Button
-                          variant="outlined"
-                          onClick={handleClickOpen}
-                          startIcon={<LocalOfferOutlinedIcon />}
-                        >
-                          Start Auction
-                        </Button>
-                      </div>
-                    </AccordionDetails>
-                  </Accordion>
-                )}
-              </div>
-            ) : (
-              <div>
-                {currentOpenAuction ? (
-                  <Accordion expanded>
-                    <AccordionSummary
-                      expandIcon={<ExpandMoreIcon />}
-                      aria-controls="panel1a-content"
-                      id="panel1a-header"
-                    >
-                      <Grid container direction="row" alignItems="center">
-                        <AccessTimeIcon style={{ marginRight: 2 }} /> There is an active auction for
-                        this NFT
-                      </Grid>
-                    </AccordionSummary>
-                    <AccordionDetails>
-                      <Typography>
-                        Sale ends on{' '}
-                        <strong>
-                          {' '}
-                          {timeConverter(
-                            BigInt(parseInt(currentOpenAuction?.sale_type?.auction?.end_date)),
-                          )}
-                        </strong>
-                      </Typography>
-                      <Typography>
-                        Token:{' '}
-                        <SymbolWithIcon
-                          symbol={
-                            currentOpenAuction?.sale_type?.auction?.config?.auction?.token?.ic
-                              ?.symbol
-                          }
-                        />
-                      </Typography>
-                      <Typography>
-                        Start price:{' '}
-                        <strong>
-                          <NatPrice
-                            value={
-                              currentOpenAuction?.sale_type?.auction?.config?.auction?.start_price
-                            }
+                  <Flex flexFlow="column">
+                    <Container size="md" padding="80px" mdPadding="16px">
+                      <Grid columns={2} mdColumns={2} gap={120} smGap={16} mdGap={40}>
+                        {odc?.hasPreviewAsset ? (
+                          <img
+                            style={{ borderRadius: '18px', width: '100%' }}
+                            src={`https://${canisterId}.raw.ic0.app/-/${params.nft_id}/preview`}
+                            onError={(e) => {
+                              e.currentTarget.src = PlaceholderImage;
+                            }}
                           />
-                        </strong>
-                      </Typography>
-                      <Typography>
-                        Minimum step:{' '}
-                        <strong>
-                          <NatPrice
-                            value={
-                              currentOpenAuction?.sale_type?.auction?.config?.auction?.min_increase
-                                ?.amount
-                            }
+                        ) : (
+                          <img
+                            style={{ borderRadius: '18px', width: '100%' }}
+                            src={PlaceholderImage}
                           />
-                        </strong>
-                      </Typography>
-                      <Typography>
-                        Highest bid:{' '}
-                        <strong>
-                          <NatPrice
-                            value={currentOpenAuction?.sale_type?.auction?.current_bid_amount}
-                          />
-                        </strong>
-                      </Typography>
-                      {currentOpenAuction?.sale_type?.auction?.config?.auction?.buy_now?.length >
-                        0 && (
-                        <Typography>
-                          Buy now:{' '}
-                          <strong>
-                            <NatPrice
-                              value={
-                                currentOpenAuction?.sale_type?.auction?.config?.auction?.buy_now[0]
-                              }
-                            />
-                          </strong>
-                        </Typography>
-                      )}
-                      <div style={{ display: 'flex', gap: 5, marginTop: 5 }}>
-                        {currentOpenAuction?.sale_type?.auction?.config?.auction?.buy_now?.length >
-                          0 && (
-                          <Button
-                            variant="contained"
-                            onClick={() => handleOpen('buyNow')}
-                            startIcon={<ShoppingCartOutlinedIcon />}
-                          >
-                            Buy now
-                          </Button>
                         )}
-                        <Button
-                          variant="outlined"
-                          onClick={() => handleOpen('bid')}
-                          startIcon={<LocalOfferOutlinedIcon />}
-                        >
-                          Make a bid
-                        </Button>
-                      </div>
-                    </AccordionDetails>
-                  </Accordion>
-                ) : (
-                  <Accordion expanded>
-                    <AccordionSummary
-                      expandIcon={<ExpandMoreIcon />}
-                      aria-controls="panel1a-content"
-                      id="panel1a-header"
-                    >
-                      <Grid container direction="row" alignItems="center">
-                        <PendingIcon style={{ marginRight: 2 }} /> This NFT is not listed for a
-                        public sale.
+                        <Flex flexFlow="column" gap={8}>
+                          <p className="secondary_color">{odc?.ownerPrincipalId}</p>
+                          <h2>
+                            <b>{odc?.displayName || odc?.id}</b>
+                          </h2>
+                          <br />
+                          <ShowMoreBlock btnText="Read More">
+                            <p className="secondary_color">{odc?.description}</p>
+                          </ShowMoreBlock>
+                          <br />
+                          <Flex gap={8} align="center">
+                            {collectionData?.hasPreviewAsset ? (
+                              <img
+                                src={`https://prptl.io/-/${canisterId}/collection/preview`}
+                                alt=""
+                                style={{ width: '32px', height: '32px', borderRadius: '7.5px' }}
+                              />
+                            ) : (
+                              <img
+                                src={PlaceholderImage}
+                                alt=""
+                                style={{ width: '32px', height: '32px', borderRadius: '7.5px' }}
+                              />
+                            )}
+                            <b>{collectionData?.displayName}</b>
+                          </Flex>
+                          <br />
+                          <HR />
+                          <Flex fullWidth justify="space-between" align="center">
+                            {odc?.auctionOpen ? (
+                              <>
+                                <Flex flexFlow="column">
+                                  <span>Current bid</span>
+                                  <strong>
+                                    <TokenIcon symbol={odc.tokenSymbol} />
+                                    {toLargerUnit(odc.currentBid, Number(odc.token.decimals))}
+                                  </strong>
+                                </Flex>
+
+                                <Flex flexFlow="column">
+                                  <span>Reserve Price</span>
+                                  <strong>
+                                    <TokenIcon symbol={odc.tokenSymbol} />
+                                    {toLargerUnit(odc.reserve, Number(odc.token.decimals))}
+                                  </strong>
+                                </Flex>
+                                {odc?.buyNow && (
+                                  <Flex flexFlow="column">
+                                    <span>Buy Now</span>
+                                    <strong>
+                                      <TokenIcon symbol={odc.tokenSymbol} />
+                                      {toLargerUnit(odc.buyNow, Number(odc.token.decimals))}
+                                    </strong>
+                                  </Flex>
+                                )}
+                              </>
+                            ) : (
+                              'Not on sale'
+                            )}
+                          </Flex>
+                          <HR />
+                          {nftEndSale && (
+                            <p className="secondary_color">
+                              {!nftEndSale || nftEndSale < currentTimeInNanos ? (
+                                <span>The sale has ended {getDiffInDays(nftEndSale)}</span>
+                              ) : (
+                                <span>{getDiffInDays(nftEndSale)}</span>
+                              )}
+                            </p>
+                          )}
+                          <br />
+                          {principalId && (
+                            <Flex gap={8} flexFlow="column">
+                              {odc?.auctionOpen ? (
+                                <>
+                                  {odc?.buyNow &&
+                                    principalId != verifyOwner &&
+                                    (nftEndSale || 9 * 1e30 > currentTimeInNanos ? (
+                                      <Button
+                                        btnType="accent"
+                                        onClick={() => onOpenEscrowModal('BuyNow')}
+                                      >
+                                        Buy Now
+                                      </Button>
+                                    ) : (
+                                      <Button disabled btnType="outlined">
+                                        Buy Now
+                                      </Button>
+                                    ))}
+
+                                  {principalId === verifyOwner ? (
+                                    odc.currentBid == 0 || odc.auctionNotStarted ? (
+                                      <Button btnType="accent" onClick={handleClickOpenEsc}>
+                                        Cancel Sale
+                                      </Button>
+                                    ) : BigInt(Number(nftEndSale || 9 * 1e30)) <
+                                      currentTimeInNanos ? (
+                                      <Button btnType="accent" onClick={handleClickOpenEsc}>
+                                        Finish Sale
+                                      </Button>
+                                    ) : (
+                                      <Button disabled btnType="outlined">
+                                        Finish Sale
+                                      </Button>
+                                    )
+                                  ) : BigInt(Number(nftEndSale || 9 * 1e30)) >
+                                    currentTimeInNanos ? (
+                                    <Button
+                                      btnType="outlined"
+                                      onClick={() => onOpenEscrowModal('Bid')}
+                                    >
+                                      Place Bid
+                                    </Button>
+                                  ) : (
+                                    <Button disabled btnType="outlined">
+                                      Place Bid
+                                    </Button>
+                                  )}
+                                </>
+                              ) : principalId === verifyOwner ? (
+                                <Button btnType="accent" onClick={handleClickOpen}>
+                                  Start an Auction
+                                </Button>
+                              ) : (
+                                <OffersPanel odc={odc} onOpenEscrowModal={onOpenEscrowModal} />
+                              )}
+                            </Flex>
+                          )}
+                        </Flex>
                       </Grid>
-                    </AccordionSummary>
-                    <AccordionDetails>
-                      You can still make an offer for this NFT by sending an escrow.
-                      <div style={{ display: 'flex', gap: 5, marginTop: 5 }}>
-                        <Button
-                          variant="outlined"
-                          onClick={handleOpen}
-                          startIcon={<LocalOfferOutlinedIcon />}
-                        >
-                          Make an offer
-                        </Button>
-                      </div>
-                    </AccordionDetails>
-                  </Accordion>
+                    </Container>
+                    <Banner bgColor="PRIMARY_1000" style={{ display: 'block' }} padding="0">
+                      <TabContent
+                        fullWidth
+                        borderBottom
+                        tabs={[
+                          { title: 'Properties', id: 'properties' },
+                          { title: 'Royalties', id: 'royalties' },
+                        ]}
+                        content={[
+                          <Container
+                            key="custom-properties"
+                            size="sm"
+                            padding="32px"
+                            smPadding="16px"
+                          >
+                            <br />
+                            <br />
+                            <br />
+                            <Flex flexFlow="column" gap={16}>
+                              {odc.displayProperties.map((p) => (
+                                <div key={p.name}>
+                                  <Grid columns={2}>
+                                    <p>{p.name.charAt(0).toUpperCase() + p.name.slice(1)}</p>
+                                    <p className="secondary_color">{p.value}</p>
+                                  </Grid>
+                                  <HR marginTop={16} />
+                                </div>
+                              ))}
+                            </Flex>
+                            <br />
+                            <br />
+                            <br />
+                          </Container>,
+                          <Container key="royalties" size="sm" padding="32px" smPadding="16px">
+                            <br />
+                            <br />
+                            <br />
+                            <Flex flexFlow="column" gap={18}>
+                              <h3>
+                                {' '}
+                                <b>Primary Royalties </b>
+                              </h3>
+                              <HR />
+                              {odc?.primaryRoyalties?.map((royalty) => (
+                                <div key={royalty.tag}>
+                                  <Grid columns={2}>
+                                    <p>{royalty.tag}</p>
+                                    <p className="secondary_color">{royalty.rate}</p>
+                                  </Grid>
+                                  <HR marginTop={18} />
+                                </div>
+                              ))}
+                              <br />
+                              <h3>
+                                {' '}
+                                <b>Secondary Royalties</b>{' '}
+                              </h3>
+                              <HR />
+                              {odc?.secondaryRoyalties?.map((royalty) => (
+                                <div key={royalty.tag}>
+                                  <Grid columns={2}>
+                                    <p>{royalty.tag}</p>
+                                    <p className="secondary_color">{royalty.rate}</p>
+                                  </Grid>
+                                  <HR marginTop={18} />
+                                </div>
+                              ))}
+                            </Flex>
+                            <br />
+                            <br />
+                            <br />
+                          </Container>,
+                        ]}
+                      />
+                    </Banner>
+                  </Flex>
                 )}
-              </div>
-            )}
-            <div hidden>
-              <Accordion expanded>
-                <AccordionSummary
-                  expandIcon={<ExpandMoreIcon />}
-                  aria-controls="panel2a-content"
-                  id="panel2a-header"
-                >
-                  <Grid container direction="row" alignItems="center">
-                    <HistoryOutlinedIcon style={{ marginRight: 2 }} /> Transaction history
-                  </Grid>
-                </AccordionSummary>
-                <AccordionDetails>There is no history for this NFT.</AccordionDetails>
-              </Accordion>
-            </div>
-          </Grid>
-        </Grid>
-        <ConfirmSalesActionModal
-          open={openConfirmation}
-          handleClose={handleClose}
-          currentToken={currentNFT}
-          action={dialogAction}
-        />
-        <StartAuctionModal
-          open={openAuction}
-          handleClose={handleClose}
-          currentToken={currentNFT?.metadata}
-        />
-      </Container>
-    </div>
+              </Flex>,
+            ]}
+            onLogOut={logout}
+            onConnect={open}
+            principal={principalId}
+          />
+
+          {openConfirmation && (
+            <ConfirmSalesActionModal
+              openConfirmation={openConfirmation}
+              handleClose={handleClose}
+              currentToken={odc?.id}
+              action={dialogAction}
+            />
+          )}
+          {openAuction && (
+            <StartAuctionModal
+              open={openAuction}
+              handleClose={handleClose}
+              onSuccess={fetchOdc}
+              currentToken={odc?.id}
+            />
+          )}
+          {openEscrowModal && (
+            <StartEscrowModal
+              open={openEscrowModal}
+              handleClose={handleCloseEscrow}
+              odc={odc}
+              escrowType={escrowType}
+              onSuccess={fetchOdc}
+            />
+          )}
+        </Flex>
+      )}
+    </>
   );
 };
